Started fleshing out BallotCrypto, which essentially is a collection of static methods that call an instance of VoteCrypto.
Deleted conditionals in ByteCrypto checking for uninitialised cryptopackages, since this will instead need to be done in BallotCrypto.

Added two new types of Vote (independent of adder's Vote class): EncryptedVote and PlaintextVote, which will be used in the new version of Ballot. Introduced the idea of typing Ballot by the type of Vote it contains (e.g. Ballot<PlaintextVote>). This should improve organisation and reduce dependence on ASEs other than for network communication. EncryptedVote will contain instances of ICiphertext, which will be dependent on the crypto method being utilised.